Implementation of advertisement service //improvements in entities, services and controllers (modified REST mappings)

+++ backend/src/main/java/com/piisw/backend/repository/LocalizationRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LocalizationRepository extends JpaRepository<Localization, Integer> {
-        Optional<Localization> findById(Integer id);
+public interface LocalizationRepository extends JpaRepository<Localization, Long> {
+        Optional<Localization> findById(Long id);
 }
